fix(icon): add possibility to update dynamically color and size
fix gravitee-io/issues#3119
