+++ package.json
@@ -11,7 +11,7 @@
     "wc/gv-*.js"
   ],
   "scripts": {
-    "build": "npm run docs && node tasks/check-i18n.js && npm run size && build-storybook",
+    "build": "npm run docs && node tasks/check-i18n.js && npm run generate:dist && npm run size && build-storybook",
     "check:i18n": "node tasks/check-i18n.js",
     "docs": "wca analyze src/** --format markdown --outDir .docs && npm run generate:custom-properties",
     "generate:custom-properties": "node tasks/generate-custom-properties.js",
@@ -37,13 +37,13 @@
   "devDependencies": {
     "@babel/core": "^7.6.3",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@storybook/addon-a11y": "^5.2.8",
-    "@storybook/addon-actions": "^5.2.8",
-    "@storybook/addon-info": "^5.2.8",
-    "@storybook/addon-knobs": "^5.2.8",
-    "@storybook/addon-notes": "^5.2.8",
-    "@storybook/html": "^5.2.8",
-    "@storybook/theming": "^5.2.8",
+    "@storybook/addon-a11y": "^5.3.9",
+    "@storybook/addon-actions": "^5.3.9",
+    "@storybook/addon-info": "^5.3.9",
+    "@storybook/addon-knobs": "^5.3.9",
+    "@storybook/addon-notes": "^5.3.9",
+    "@storybook/html": "^5.3.9",
+    "@storybook/theming": "^5.3.9",
     "babel-loader": "^8.0.6",
     "babel-plugin-template-html-minifier": "^3.1.0",
     "bundle-phobia-cli": "^0.14.1",
