+++ src/icons/shapes/thirdparty.js
@@ -1,3 +1,3 @@
-export const ThirdpartyShapes = { github: '<svg viewBox="0 0 32 32"><path fill-rule="evenodd" clip-rule="evenodd" fill="#000" d="M16 4.297c-6.63 0-12 5.373-12 12 0 5.303 3.438 9.8 8.205 11.385.6.113.82-.258.82-.577 0-.285-.01-1.04-.015-2.04-3.338.724-4.042-1.61-4.042-1.61C8.422 22.07 7.633 21.7 7.633 21.7c-1.087-.744.084-.729.084-.729 1.205.084 1.838 1.236 1.838 1.236 1.07 1.835 2.809 1.305 3.495.998.108-.776.417-1.305.76-1.605-2.665-.3-5.466-1.332-5.466-5.93 0-1.31.465-2.38 1.235-3.22-.135-.303-.54-1.523.105-3.176 0 0 1.005-.322 3.3 1.23.96-.267 1.98-.399 3-.405 1.02.006 2.04.138 3 .405 2.28-1.552 3.285-1.23 3.285-1.23.645 1.653.24 2.873.12 3.176.765.84 1.23 1.91 1.23 3.22 0 4.61-2.805 5.625-5.475 5.92.42.36.81 1.096.81 2.22 0 1.606-.015 2.896-.015 3.286 0 .315.21.69.825.57C24.565 26.092 28 21.592 28 16.297c0-6.627-5.373-12-12-12"/>', google: '<svg viewBox="0 0 833.5 844.3"><path fill="#fff" d="M100 100h633.5v644.3H100z"/><path d="M683.5 428.4c0-18.5-1.5-37.1-4.7-55.3H422.1v104.8h147c-6.1 33.8-25.7 63.7-54.4 82.7v68h87.7c51.5-47.4 81.1-117.4 81.1-200.2z" fill="#4285f4"/><path d="M422.1 694.3c73.4 0 135.3-24.1 180.4-65.7l-87.7-68c-24.4 16.6-55.9 26-92.6 26-71 0-131.2-47.9-152.8-112.3h-90.5v70.1c46.2 91.9 140.3 149.9 243.2 149.9z" fill="#34a853"/><path d="M269.3 474.3c-11.4-33.8-11.4-70.4 0-104.2V300h-90.4c-38.6 76.9-38.6 167.5 0 244.4l90.4-70.1z" fill="#fbbc04"/><path d="M422.1 257.7c38.8-.6 76.3 14 104.4 40.8l77.7-77.7C555 174.6 489.7 149.2 422.1 150c-102.9 0-197 58-243.2 150l90.4 70.1c21.5-64.5 81.8-112.4 152.8-112.4z" fill="#ea4335"/>', graviteeio_am: '<svg viewBox="0 0 115 115"><path d="M81.96 19.04c-5.69 1.8-12.52 3.5-18.49 6.11L86 27.97l-4.04-8.93zM60.39 91.42l-11.64-5.56a208.505 208.505 0 01-17.22-9.94l-.78-.45v-.05c-6.34-4.12-12.19-8.45-17.39-12.86 1.21 15.56 10.75 29.31 23.56 36.72 6.65 3.71 14.3 5.82 22.46 5.82 8.15 0 15.81-2.12 22.46-5.82-6.89-2-14.12-4.64-21.45-7.86z" fill="#2B2E31"/><path d="M30.75 43.08c-6.36 4.15-12.23 8.51-17.43 12.96a46.642 46.642 0 000 5.82c5.2 4.44 11.07 8.81 17.43 12.96V43.08zM83.23 60.51l4.74 14.94.03.03-.79.45c-4.52 2.91-9.29 5.72-14.23 8.37-.99.53-1.98 1.04-2.96 1.55l-9.09 5.25c7.37 3.22 14.63 5.85 21.52 7.82 12.82-7.41 21.74-20.81 22.94-36.37.06-.96.06.02.06-.96l-22.22-1.08z" fill="#2B2E31"/><path d="M77.45 16.46a46.08 46.08 0 00-18.07-3.67c-8.15 0-16.43 2.47-23.08 6.18-12.82 7.41-21.74 20.81-22.94 36.37 5.2-4.41 11.05-8.74 17.39-12.86v-.06l.76-.44c4.53-2.92 9.3-5.73 14.26-8.39.99-.54 1.99-1.05 2.98-1.57l11.63-5.55c6.11-2.68 15.62-6.02 21.45-7.86l-4.38-2.15z" fill="#2B2E31"/><path fill="#32ABDF" d="M65.53 56.19a4.87 4.87 0 11-9.739.001 4.87 4.87 0 019.739-.001z"/>', oidc: '<svg viewBox="0 0 90 90"><g fill="#000"><path d="M41.928 15.65v63.5l11.289-5.313V10.135zM81.023 34.277l1.559 16.18-21.831-4.75"/><path d="M19.35 54.982c0-7.999 8.716-14.735 20.596-16.791v-7.17C21.775 33.217 8.061 43.109 8.061 54.982c0 12.301 14.72 22.475 33.867 24.168v-7.077C29.046 70.458 19.35 63.42 19.35 54.983m35.848-23.96v7.168c4.727.818 8.956 2.373 12.33 4.459l7.983-4.933c-5.44-3.363-12.461-5.745-20.313-6.694"/></g>' };
+export const ThirdpartyShapes = { github: '<svg viewBox="0 0 32 32"><path fill-rule="evenodd" clip-rule="evenodd" fill="#000" d="M16 4.297c-6.63 0-12 5.373-12 12 0 5.303 3.438 9.8 8.205 11.385.6.113.82-.258.82-.577 0-.285-.01-1.04-.015-2.04-3.338.724-4.042-1.61-4.042-1.61C8.422 22.07 7.633 21.7 7.633 21.7c-1.087-.744.084-.729.084-.729 1.205.084 1.838 1.236 1.838 1.236 1.07 1.835 2.809 1.305 3.495.998.108-.776.417-1.305.76-1.605-2.665-.3-5.466-1.332-5.466-5.93 0-1.31.465-2.38 1.235-3.22-.135-.303-.54-1.523.105-3.176 0 0 1.005-.322 3.3 1.23.96-.267 1.98-.399 3-.405 1.02.006 2.04.138 3 .405 2.28-1.552 3.285-1.23 3.285-1.23.645 1.653.24 2.873.12 3.176.765.84 1.23 1.91 1.23 3.22 0 4.61-2.805 5.625-5.475 5.92.42.36.81 1.096.81 2.22 0 1.606-.015 2.896-.015 3.286 0 .315.21.69.825.57C24.565 26.092 28 21.592 28 16.297c0-6.627-5.373-12-12-12"/>', google: '<svg class="no-color" viewBox="0 0 833.5 844.3"><path fill="#fff" d="M100 100h633.5v644.3H100z"/><path d="M683.5 428.4c0-18.5-1.5-37.1-4.7-55.3H422.1v104.8h147c-6.1 33.8-25.7 63.7-54.4 82.7v68h87.7c51.5-47.4 81.1-117.4 81.1-200.2z" fill="#4285f4"/><path d="M422.1 694.3c73.4 0 135.3-24.1 180.4-65.7l-87.7-68c-24.4 16.6-55.9 26-92.6 26-71 0-131.2-47.9-152.8-112.3h-90.5v70.1c46.2 91.9 140.3 149.9 243.2 149.9z" fill="#34a853"/><path d="M269.3 474.3c-11.4-33.8-11.4-70.4 0-104.2V300h-90.4c-38.6 76.9-38.6 167.5 0 244.4l90.4-70.1z" fill="#fbbc04"/><path d="M422.1 257.7c38.8-.6 76.3 14 104.4 40.8l77.7-77.7C555 174.6 489.7 149.2 422.1 150c-102.9 0-197 58-243.2 150l90.4 70.1c21.5-64.5 81.8-112.4 152.8-112.4z" fill="#ea4335"/>', graviteeio_am: '<svg class="no-color" viewBox="0 0 115 115"><path d="M81.96 19.04c-5.69 1.8-12.52 3.5-18.49 6.11L86 27.97l-4.04-8.93zM60.39 91.42l-11.64-5.56a208.505 208.505 0 01-17.22-9.94l-.78-.45v-.05c-6.34-4.12-12.19-8.45-17.39-12.86 1.21 15.56 10.75 29.31 23.56 36.72 6.65 3.71 14.3 5.82 22.46 5.82 8.15 0 15.81-2.12 22.46-5.82-6.89-2-14.12-4.64-21.45-7.86z" fill="#2B2E31"/><path d="M30.75 43.08c-6.36 4.15-12.23 8.51-17.43 12.96a46.642 46.642 0 000 5.82c5.2 4.44 11.07 8.81 17.43 12.96V43.08zM83.23 60.51l4.74 14.94.03.03-.79.45c-4.52 2.91-9.29 5.72-14.23 8.37-.99.53-1.98 1.04-2.96 1.55l-9.09 5.25c7.37 3.22 14.63 5.85 21.52 7.82 12.82-7.41 21.74-20.81 22.94-36.37.06-.96.06.02.06-.96l-22.22-1.08z" fill="#2B2E31"/><path d="M77.45 16.46a46.08 46.08 0 00-18.07-3.67c-8.15 0-16.43 2.47-23.08 6.18-12.82 7.41-21.74 20.81-22.94 36.37 5.2-4.41 11.05-8.74 17.39-12.86v-.06l.76-.44c4.53-2.92 9.3-5.73 14.26-8.39.99-.54 1.99-1.05 2.98-1.57l11.63-5.55c6.11-2.68 15.62-6.02 21.45-7.86l-4.38-2.15z" fill="#2B2E31"/><path fill="#32ABDF" d="M65.53 56.19a4.87 4.87 0 11-9.739.001 4.87 4.87 0 019.739-.001z"/>', oidc: '<svg viewBox="0 0 90 90"><g fill="#000"><path d="M41.928 15.65v63.5l11.289-5.313V10.135zM81.023 34.277l1.559 16.18-21.831-4.75"/><path d="M19.35 54.982c0-7.999 8.716-14.735 20.596-16.791v-7.17C21.775 33.217 8.061 43.109 8.061 54.982c0 12.301 14.72 22.475 33.867 24.168v-7.077C29.046 70.458 19.35 63.42 19.35 54.983m35.848-23.96v7.168c4.727.818 8.956 2.373 12.33 4.459l7.983-4.933c-5.44-3.363-12.461-5.745-20.313-6.694"/></g>' };
 window.GvIcons = window.GvIcons || {};
 window.GvIcons.thirdparty = ThirdpartyShapes;
