--- wc/gv-icons.js
@@ -1,1 +1,0 @@
-import '../src/icons/gv-icons';
